backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Azure AI and OpenAI
azure-identity>=1.15.0
openai>=1.0.0
httpx>=0.24.0
